requirements.txt
blabel==0.1.6
pandas==2.2.3
Pillow==11.0.0
PyDrive2==1.20.0
PyYAML==6.0.2
segno==1.6.1
streamlit==1.39.0
streamlit_authenticator==0.4.1
supabase==2.9.1
